refactor: Add events history retrieval to get_user method
Update the `get_user` method in `compute.py` to include the retrieval of events history for a user. This is achieved by calling the `get_events_history` method from the `Common` class and passing the appropriate query. The events are then added to the user object. This enhancement provides additional information about the user's events and improves the completeness of the user data.