```text Refactor Compute class to update user and send feedback message
This commit refactors the `Compute` class in the `call_webhook/compute.py` file. It introduces two new methods: `update_user` and `send_feedback_message`.

The `update_user` method now takes additional parameters `expert` and `user` and updates the user's `isBusy` status. If the expert is not internal, it also updates the user's `numberOfCalls` by decrementing it by 1 if it is greater than 0.

The `send_feedback_message` method now takes additional parameters `expert` and `user` and sends a feedback message if both the user and expert are found.

These changes improve the functionality and maintainability of the `Compute` class.
```<|diff_marker|> --- git-commit-message/git-commit-message.py
<|diff_marker|> 1000
+import difflib
<|diff_marker|> --- git-commit-message/git-commit-message.py
<|diff_marker|> 1001
+import re
<|diff_marker|> --- git-commit-message/git-commit-message.py
<|diff_marker|> 1020
+
+def remove_meta_information(commit_message):
+    # Remove issue references
+    commit_message = re.sub(r'#\d+', '', commit_message)
+    # Remove tags
+    commit_message = re.sub(r'\[.*?\]', '', commit_message)
+    # Remove author names
+    commit_message = re.sub(r'@.*?\b', '', commit_message)
+    return commit_message.strip()
<|diff_marker|> --- git-commit-message/git-commit-message.py
<|diff_marker|> 1039
+
+commit_message = """
+Refactor Compute class to update user and send feedback message

This commit refactors the `Compute` class in the `call_webhook/compute.py` file. It introduces two new methods: `update_user` and `send_feedback_message`.

The `update_user` method now takes additional parameters `expert` and `user` and updates the user's `isBusy` status. If the expert is not internal, it also updates the user's `numberOfCalls` by decrementing it by 1 if it is greater than 0.

The `send_feedback_message` method now takes additional parameters `expert` and `user` and sends a feedback message if both the user and expert are found.

These changes improve the functionality and maintainability of the `Compute` class.
+"""
+
+commit_message = remove_meta_information(commit_message)
+print("```text\n" + commit_message + "\n```")